Split repository and service classes into separate files
