+++ Parkman/Infrastructure/Repositories/Entities/ParkingSpotRepository.cs
@@ -1,0 +1,12 @@
+using Microsoft.Extensions.Logging;
+using Parkman.Domain.Entities;
+
+namespace Parkman.Infrastructure.Repositories.Entities;
+
+public class ParkingSpotRepository : GenericRepository<ParkingSpot>, IParkingSpotRepository
+{
+    public ParkingSpotRepository(
+        ApplicationDbContext context,
+        ILogger<GenericRepository<ParkingSpot>> logger)
+        : base(context, logger) { }
+}
