+++ Parkman/Infrastructure/Repositories/Entities/ProfileReservationRepository.cs
@@ -1,0 +1,12 @@
+using Microsoft.Extensions.Logging;
+using Parkman.Domain.Entities;
+
+namespace Parkman.Infrastructure.Repositories.Entities;
+
+public class ProfileReservationRepository : GenericRepository<ProfileReservation>, IProfileReservationRepository
+{
+    public ProfileReservationRepository(
+        ApplicationDbContext context,
+        ILogger<GenericRepository<ProfileReservation>> logger)
+        : base(context, logger) { }
+}
