+++ Parkman/Infrastructure/Services/Entities/ICompanyProfileService.cs
@@ -5,7 +5,3 @@ namespace Parkman.Infrastructure.Services.Entities;
 
 public interface ICompanyProfileService : IGenericService<CompanyProfile> { }
 
-public class CompanyProfileService : GenericService<CompanyProfile>, ICompanyProfileService
-{
-    public CompanyProfileService(ICompanyProfileRepository repository) : base(repository) { }
-}
