+++ Parkman/Infrastructure/Services/Entities/ICompanyReservationService.cs
@@ -5,7 +5,3 @@ namespace Parkman.Infrastructure.Services.Entities;
 
 public interface ICompanyReservationService : IGenericService<CompanyReservation> { }
 
-public class CompanyReservationService : GenericService<CompanyReservation>, ICompanyReservationService
-{
-    public CompanyReservationService(ICompanyReservationRepository repository) : base(repository) { }
-}
