+++ Parkman/Infrastructure/Services/Entities/IParkingLotService.cs
@@ -5,7 +5,3 @@ namespace Parkman.Infrastructure.Services.Entities;
 
 public interface IParkingLotService : IGenericService<ParkingLot> { }
 
-public class ParkingLotService : GenericService<ParkingLot>, IParkingLotService
-{
-    public ParkingLotService(IParkingLotRepository repository) : base(repository) { }
-}
