+++ Parkman/Infrastructure/Services/Entities/IParkingSpotService.cs
@@ -5,7 +5,3 @@ namespace Parkman.Infrastructure.Services.Entities;
 
 public interface IParkingSpotService : IGenericService<ParkingSpot> { }
 
-public class ParkingSpotService : GenericService<ParkingSpot>, IParkingSpotService
-{
-    public ParkingSpotService(IParkingSpotRepository repository) : base(repository) { }
-}
