+++ Parkman/Infrastructure/Services/Entities/IPersonProfileService.cs
@@ -5,7 +5,3 @@ namespace Parkman.Infrastructure.Services.Entities;
 
 public interface IPersonProfileService : IGenericService<PersonProfile> { }
 
-public class PersonProfileService : GenericService<PersonProfile>, IPersonProfileService
-{
-    public PersonProfileService(IPersonProfileRepository repository) : base(repository) { }
-}
