+++ Parkman/Infrastructure/Services/Entities/IProfileReservationService.cs
@@ -5,7 +5,3 @@ namespace Parkman.Infrastructure.Services.Entities;
 
 public interface IProfileReservationService : IGenericService<ProfileReservation> { }
 
-public class ProfileReservationService : GenericService<ProfileReservation>, IProfileReservationService
-{
-    public ProfileReservationService(IProfileReservationRepository repository) : base(repository) { }
-}
