+++ Parkman/Infrastructure/Services/Entities/IReservationService.cs
@@ -5,7 +5,3 @@ namespace Parkman.Infrastructure.Services.Entities;
 
 public interface IReservationService : IGenericService<Reservation> { }
 
-public class ReservationService : GenericService<Reservation>, IReservationService
-{
-    public ReservationService(IReservationRepository repository) : base(repository) { }
-}
