+++ Parkman/Infrastructure/Services/Entities/IVehicleService.cs
@@ -5,7 +5,3 @@ namespace Parkman.Infrastructure.Services.Entities;
 
 public interface IVehicleService : IGenericService<Vehicle> { }
 
-public class VehicleService : GenericService<Vehicle>, IVehicleService
-{
-    public VehicleService(IVehicleRepository repository) : base(repository) { }
-}
