+++ Parkman/Infrastructure/Services/Entities/ParkingLotService.cs
@@ -1,0 +1,9 @@
+using Parkman.Domain.Entities;
+using Parkman.Infrastructure.Repositories.Entities;
+
+namespace Parkman.Infrastructure.Services.Entities;
+
+public class ParkingLotService : GenericService<ParkingLot>, IParkingLotService
+{
+    public ParkingLotService(IParkingLotRepository repository) : base(repository) { }
+}
