+++ Parkman/Infrastructure/Services/Entities/PersonProfileService.cs
@@ -1,0 +1,9 @@
+using Parkman.Domain.Entities;
+using Parkman.Infrastructure.Repositories.Entities;
+
+namespace Parkman.Infrastructure.Services.Entities;
+
+public class PersonProfileService : GenericService<PersonProfile>, IPersonProfileService
+{
+    public PersonProfileService(IPersonProfileRepository repository) : base(repository) { }
+}
