+++ Parkman/Infrastructure/Services/Entities/ProfileReservationService.cs
@@ -1,0 +1,9 @@
+using Parkman.Domain.Entities;
+using Parkman.Infrastructure.Repositories.Entities;
+
+namespace Parkman.Infrastructure.Services.Entities;
+
+public class ProfileReservationService : GenericService<ProfileReservation>, IProfileReservationService
+{
+    public ProfileReservationService(IProfileReservationRepository repository) : base(repository) { }
+}
