+++ Parkman/Infrastructure/Services/Entities/VehicleService.cs
@@ -1,0 +1,9 @@
+using Parkman.Domain.Entities;
+using Parkman.Infrastructure.Repositories.Entities;
+
+namespace Parkman.Infrastructure.Services.Entities;
+
+public class VehicleService : GenericService<Vehicle>, IVehicleService
+{
+    public VehicleService(IVehicleRepository repository) : base(repository) { }
+}
